Move traceid spanid traceflags to common
