+++ opentelemetry/src/trace/mod.rs
@@ -182,11 +182,12 @@ pub use self::{
         get_active_span, mark_span_as_active, FutureExt, SpanRef, TraceContextExt, WithContext,
     },
     span::{Span, SpanKind, Status},
-    span_context::{SpanContext, SpanId, TraceFlags, TraceId, TraceState},
+    span_context::{SpanContext, TraceState},
     tracer::{SamplingDecision, SamplingResult, SpanBuilder, Tracer},
     tracer_provider::TracerProvider,
 };
 use crate::KeyValue;
+pub use crate::{SpanId, TraceFlags, TraceId};
 use std::sync::PoisonError;
 
 // TODO - Move ExportError and TraceError to opentelemetry-sdk
